9th

+++ aloeda-service/install.sh
@@ -3,6 +3,17 @@
 # Author      : Adeola Ojo
 # Description : Simple app service installer
 
+# vars
+START_CONTAINERS=false
+
+while getopts s: option
+do
+    case "${option}"
+    in
+    s) START_CONTAINERS=${OPTARG};;
+    esac
+done
+
 # build
 ./gradlew clean build
 
@@ -12,10 +23,14 @@ docker image prune -f
 docker volume prune -f
 docker network prune -f
 
-# switch to installer
-cd installer
+if [[ "$START_CONTAINERS" == 'true' ]]; then
+    echo "starting apps $START_CONTAINERS"
 
-# install
-docker-compose rm -f
-docker-compose pull
-docker-compose up --build -d
+    # switch to installer
+    cd installer
+
+    # install
+    docker-compose rm -f
+    docker-compose pull
+    docker-compose up --build -d
+fi
